Year 2023, Day 24, Without python Z3, Just COS
Made a solution with COS and without using Z3. I already made for part 1 the equations for the solutions so now just running on small range (defined by half of maxdelta that was found in the data, worked for me on 3 real data and one example data) and filtering out the one that there's no solution or it's not an integer.